Membership: move Exceptions to dedicated files.
More idiomatic, and avoids potential future circular import issues.
