--- borrowd_groups/__init__.py
@@ -1,5 +1,0 @@
-from borrowd import BorrowdException
-
-
-class ExistingMemberException(BorrowdException):
-    pass
