+++ borrowd_groups/exceptions.py
@@ -1,0 +1,5 @@
+from borrowd.exceptions import BorrowdException
+
+
+class ExistingMemberException(BorrowdException):
+    pass
